pcaw/pcaw_backend/requirements.txt
asgiref==3.6.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
coverage==7.2.3
cryptography==40.0.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0
django-cors-headers==3.14.0
django-extensions==3.2.1
django-oauth-toolkit==2.1.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
django-environ==0.10.0
drf-social-oauth2==1.3.0
ecdsa==0.18.0
fuzzywuzzy==0.18.0
google-api-core==2.11.0
google-auth==2.17.2
google-cloud-core==2.3.2
google-cloud-secret-manager==2.16.1
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
graphviz==0.20.1
idna==3.4
jwcrypto==1.4.2
Levenshtein==0.20.9
numpy==1.24.2
oauthlib==3.2.2
pandas==2.0.0
Pillow==9.5.0
protobuf==4.22.1
psycopg2==2.9.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydot==1.4.2
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.8
python-jose==3.3.0
python-Levenshtein==0.20.9
python3-openid==3.2.0
pytz==2023.3
rapidfuzz==2.15.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.1
sqlparse==0.4.3
tzdata==2023.3
urllib3==1.26.15
wrapt==1.15.0
django-storages==1.13.2
gunicorn>=20.1.0,<20.2
et-xmlfile==1.1.0
openpyxl==3.1.2
odfpy==1.4.1
pyxlsb==1.0.10
xlrd==2.0.1 
xlsxwriter==3.1.0